mcq-backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
pydantic
python-jose
requests
python-multipart
typing-extensions
asyncpg
anyio
httpx
google-genai
openai
pinecone
huggingface-hub
pdfplumber
numpy
bcrypt
